+++ app/controllers/users_controller.rb
@@ -1,0 +1,23 @@
+class UsersController < ApplicationController
+  before_action :set_user, only: [:show]
+
+  # GET /users
+  # GET /users.json
+  def index
+    @users = User.all
+  end
+
+  # GET /users/1
+  # GET /users/1.json
+  def show
+  end
+
+  # TODO: add edit, create, update, and destroy. Then make the view and specs to test that we can add and edit users.
+
+  private
+
+  # Use callbacks to share common setup or constraints between actions.
+  def set_user
+    @user = User.find(params[:id])
+  end
+end
