+++ app/models/email.rb
@@ -1,6 +1,4 @@
 class Email < ApplicationRecord
-  # allow this model to be used as a contact
-  include OwnerContactMap
   
   validates :value, presence: true, format: { with: /\A[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}\z/i }
