+++ app/models/phone.rb
@@ -1,6 +1,4 @@
 class Phone < ApplicationRecord
-  # allow this model to be owned as a contact
-  include OwnerContactMap
 
   validates :value, presence: true, uniqueness: true
