+++ app/models/program.rb
@@ -1,6 +1,6 @@
 class Program < ApplicationRecord
   has_many :program_memberships
-  has_many :people, through: :program_memberships
+  has_many :users, through: :program_memberships
   has_many :roles, through: :program_memberships
 
   validates :name, presence: true, uniqueness: {case_sensitive: false}
