+++ app/models/program_membership.rb
@@ -1,5 +1,5 @@
 class ProgramMembership < ApplicationRecord
-  belongs_to :person
+  belongs_to :user
   belongs_to :program
   belongs_to :role
