+++ app/models/role.rb
@@ -1,6 +1,6 @@
 class Role < ApplicationRecord
   has_many :program_memberships
-  has_many :people, through: :program_memberships
+  has_many :users, through: :program_memberships
   has_many :programs, through: :program_memberships
   
   validates :name, presence: true, uniqueness: {case_sensitive: false}
