--- app/views/people/index.json.jbuilder
@@ -1,1 +1,0 @@
-json.people @people.map(&:full_name)
