--- app/views/people/show.html.erb
@@ -1,66 +1,0 @@
-<% content_for :headline do %>
-  <%= @person.full_name %>
-<% end %>
-
-<% content_for(:headline_buttons) do %>
-  <%= link_to 'Back', postal_codes_path, class: 'btn btn-sm btn-outline-secondary' %>
-  <%= link_to 'JSON', person_path(@person, format: 'json'), class: 'btn btn-sm btn-secondary', target: '_blank' %>
-<% end %>
-
-<div class="row">
-  <div class="col-sm-4">
-    <div class="card">
-      <div class="card-header">
-        <span data-feather="phone"></span>
-        Contact Info
-      </div>
-        
-      <ul class="list-group list-group-flush">
-        <% @person.phones.each do |phone| %>
-          <li class="list-group-item"><%= phone.value %></li>
-        <% end %>
-
-        <% @person.emails.each do |email| %>
-          <li class="list-group-item"><%= link_to email.value, "mailto:#{email.value}" %></li>
-        <% end %>
-
-        <% @person.addresses.each do |address| %>
-          <li class="list-group-item">
-            <%= address.line1 %><br>
-          
-            <% if address.line2 %>
-              <%= address.line2 %><br>
-            <% end %>
-          
-            <%= address.city %>, <%= address.state %> <%= address.zip %>
-          </li>
-        <% end %>
-      </ul>
-    </div> <!-- card -->
-  </div>
-  
-  <div class="col-sm-4">
-    <div class="card">
-      <div class="card-header">
-        <span data-feather="shield"></span>
-        Program Memberships
-      </div>
-        
-      <ul class="list-group list-group-flush">
-        <% if @person.program_memberships.empty? %>
-          <li class="list-group-item">None at this time.</li>
-        <% else %>
-          <% @person.program_memberships.each do |membership| %>
-            <li class="list-group-item">
-              <b><%= membership.program.name %> <%= membership.role.name %></b>
-              <br>
-              
-              <%= membership.start_date.strftime('%Y-%m-%d') %> to
-              <%= membership.end_date ? membership.end_date.strftime('%Y-%m-%d') : 'Present' %>
-            </li>
-          <% end %>
-        <% end %>
-      </ul>
-    </div> <!-- card -->
-  </div>
-</div> <!-- row -->
