+++ app/views/users/index.json.jbuilder
@@ -1,0 +1,1 @@
+json.users @users.map(&:full_name)
