+++ app/views/users/show.html.erb
@@ -1,0 +1,35 @@
+<% content_for :headline do %>
+  <%= @user.full_name %>
+<% end %>
+
+<% content_for(:headline_buttons) do %>
+  <%= link_to 'Back', postal_codes_path, class: 'btn btn-sm btn-outline-secondary' %>
+  <%= link_to 'JSON', user_path(@user, format: 'json'), class: 'btn btn-sm btn-secondary', target: '_blank' %>
+<% end %>
+
+<div class="row">
+  <div class="col-sm-4">
+    <div class="card">
+      <div class="card-header">
+        <span data-feather="shield"></span>
+        Program Memberships
+      </div>
+        
+      <ul class="list-group list-group-flush">
+        <% if @user.program_memberships.empty? %>
+          <li class="list-group-item">None at this time.</li>
+        <% else %>
+          <% @user.program_memberships.each do |membership| %>
+            <li class="list-group-item">
+              <b><%= membership.program.name %> <%= membership.role.name %></b>
+              <br>
+              
+              <%= membership.start_date.strftime('%Y-%m-%d') %> to
+              <%= membership.end_date ? membership.end_date.strftime('%Y-%m-%d') : 'Present' %>
+            </li>
+          <% end %>
+        <% end %>
+      </ul>
+    </div> <!-- card -->
+  </div>
+</div> <!-- row -->
