+++ db/migrate/20200123163752_change_program_memberships_from_person_to_user.rb
@@ -1,0 +1,7 @@
+class ChangeProgramMembershipsFromPersonToUser < ActiveRecord::Migration[6.0]
+  def change
+    change_table :program_memberships do |t|
+      t.rename :person_id, :user_id
+    end
+  end
+end
