--- spec/factories/contact_owners.rb
@@ -1,6 +1,0 @@
-FactoryBot.define do
-  factory :contact_owner do
-    association :owner, factory: :person
-    association :contact, factory: :phone
-  end
-end
