+++ spec/factories/program_memberships.rb
@@ -1,6 +1,6 @@
 FactoryBot.define do
   factory :program_membership do
-    association :person
+    association :user
     association :program
     association :role
   end
