+++ spec/models/email_spec.rb
@@ -1,7 +1,6 @@
 require 'rails_helper'
 
 RSpec.describe Email, type: :model do
-  it_behaves_like 'contact', :email
   
   #############
   # Validations
