+++ spec/models/phone_spec.rb
@@ -1,7 +1,6 @@
 require 'rails_helper'
 
 RSpec.describe Phone, type: :model do
-  it_behaves_like "contact", :phone
   
   #############
   # Validations
