+++ spec/models/program_membership_spec.rb
@@ -5,7 +5,7 @@ RSpec.describe ProgramMembership, type: :model do
   # Associations
   ##############
 
-  it { should belong_to :person }
+  it { should belong_to :user }
   it { should belong_to :program }
   it { should belong_to :role }
