+++ spec/routing/people_routing_spec.rb
@@ -1,13 +1,13 @@
 require "rails_helper"
 
-RSpec.describe PeopleController, type: :routing do
+RSpec.describe UsersController, type: :routing do
   describe "routing" do
     it "routes to #index" do
-      expect(:get => "/people").to route_to("people#index")
+      expect(:get => "/users").to route_to("users#index")
     end
 
     it "routes to #show" do
-      expect(get: '/people/1').to route_to("people#show", id: '1')
+      expect(get: '/users/1').to route_to("users#show", id: '1')
     end
   end
 end
